Allow local rbd user and secret_uuid configuration
By default, the rbd_user and rbd_secret_uuid are specified in the
nova-volume/cinder configuration and passed to nova-compute when
volumes are attached to instances.

This change allows these values to be specified locally in
nova-compute which means access control to RADOS devices in ceph
can be managed independently from nova-volume/cinder with no
requirement for consistent uuid's for libvirt secrets.

Fixes bug 1065883.

Change-Id: I9f07d040ae267bfbe8f794a5d22d327106314cc6