Fixed HyperV to get disk stats of instances drive
It fixes bug 1077922

The current version of the HyperV Compute driver was returning the disk stats about the C: drive always.

This patch makes HyperV driver extract the current drive from the flag 'instances_path'.

In order to optimize the procedure, the disk stats extraction is also unified in one method.

Change-Id: I4920cf2cc07d0f514c0358144b7bceb15d439a63