Remove uses of instance['instance_type'] from libvirt driver.
Note that this includes a few tweaks to the libvirt tests where the instance
type in use was expected to have no swap, but the actual id-matching fake
did. This integration makes it a little harder for the tests to be out
of whack like that, which is good.

This is one change in a series aimed at removing the use of instance-linked
instance_type objects, in favor of the decoupled type data in
system_metadata. See bug 1140119 for more details.

Change-Id: I190258bd0947944d9bcdf49956eed40d402606f5