Remove instance['instance_type'] relationship from db api
This removes the entire relationship between an instance and the flavor
it was created from. As such, it removes two tests entirely which
examined the old behavior.

This is one change in a series aimed at removing the use of instance-linked
instance_type objects, in favor of the decoupled type data in
system_metadata. See bug 1140119 for more details.

Change-Id: I8b525a900bf2dc6b193e54aae35fc32248e4cff4