Add cpuset attr to vcpu conf in libvirt xml
Currently the instances can use all of the pcpu of compute node, the
host may become slow when vcpus of instances are busy, so we need to
pin vcpus to the specific pcpus of host instead of all pcpus.
Also added a cache of total vcpus num in libvirt driver here.

DocImpact: please refer to the bp page.
Implements blueprint: instance-vcpu-pin
Change-Id: I6b6e805c15b249cc5e7517e26619d79bb0183de8