Remove multi scheduler.
This patch removes the multi scheduler.  It doesn't provide any value.
The only code that exists here is for passing through requests to the
compute scheduler.  There is nothing that uses the 'default' scheduler.

The only case where this code could have been useful is if it was
patched out of tree.  At that point, it seems easier to just provide
your own scheduler driver that does what you want instead of trying to
patch this one, IMO.

DocImpact: The multi scheduler no longer exists.

Change-Id: I30902d1ce0aa5baef6693dc4a3b1057cccb27841