New select_destinations scheduler call
select_destinations returns a list of (host, node) tuples that satisfy
the request_spec and filter_properties passed to it.  This will allow
the conductor to get a list of destinations for workflows such as
creating or resizing an instance and then handle the orchestration
itself.

part of bp query-scheduler

Change-Id: I1a42cea64dbad67562d7efe8d759e6efb5ec8121